requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sentence-transformers>=2.7.0
faiss-cpu==1.8.0
PyPDF2==3.0.1
python-docx==1.1.0
numpy<2.0.0,>=1.24.3
pandas>=2.0.3
requests>=2.31.0
python-dotenv>=1.0.0
jinja2>=3.1.2
aiofiles>=23.2.1
